requirements.txt
janis-pipelines[bioinformatics]